Move AutoMode to a new file
