+++ src/nextline_schedule/auto/callback.py
@@ -1,5 +1,4 @@
 import asyncio
-from collections.abc import AsyncIterator
 from logging import getLogger
 from typing import Any, Callable, Coroutine
 
@@ -9,35 +8,6 @@ from nextline.plugin.spec import Context, hookimpl
 from nextline_schedule.types import Statement
 
 from .state_machine.machine import AutoModeStateMachine
-
-
-class AutoMode:
-    def __init__(
-        self,
-        nextline: Nextline,
-        pull_func: Callable[[], Coroutine[Any, Any, Statement]],
-    ):
-        self._machine = build_state_machine(nextline=nextline, pull_func=pull_func)
-
-    @property
-    def state(self) -> str:
-        return self._machine.state
-
-    def subscribe_state(self) -> AsyncIterator[str]:
-        return self._machine.subscribe_state()
-
-    async def turn_on(self) -> None:
-        await self._machine.turn_on()  # type: ignore
-
-    async def turn_off(self) -> None:
-        await self._machine.turn_off()  # type: ignore
-
-    async def __aenter__(self) -> 'AutoMode':
-        await self._machine.__aenter__()
-        return self
-
-    async def __aexit__(self, exc_type, exc_value, traceback) -> None:
-        await self._machine.__aexit__(exc_type, exc_value, traceback)
 
 
 def build_state_machine(
